refactor: accordion and ua analytics last 30 days
